Add ability to autogenerate & serve LetsEncrypt SSL/TLS Certificate (#4)
This change uses the autocert package to add initial support for generating and serving a real live SSL certificate from LetsEncrypt via `ssl-proxy`. Now, serving your application with SSL is as easy as:

```sh
./ssl-proxy -from=0.0.0.0:443 -to=localhost:8080 -domain=mydomain.com
```

this will attempt to fetch a certificate for mydomain.com from LetsEncrypt (port 443 must be bindable by this process) and then begin proxying HTTPS traffic on :443 to any server running on :8080. Easy. This closes #1.