Update Docker build to use Go 1.13.7 (#26)
This updates the Docker image to use Go 1.13.7, as Go 1.13+ is now required after #22. The GitHub actions CI was already using Go 1.13 so this was not caught there. If we will continue to maintain a Docker image to build ssl-proxy, some work should go into testing this docker build using GitHub actions as well!

Separately, we should investigate building against Go 1.14.

This closes #25.